+++ runtime/Cargo.toml
@@ -47,6 +47,8 @@ tiny-keccak = "1.4.2"
 substrate-keyring = { git = "https://github.com/paritytech/substrate", branch = "polkadot-master" }
 substrate-trie = { git = "https://github.com/paritytech/substrate", branch = "polkadot-master" }
 trie-db = "0.12"
+lazy_static = "1.3"
+parking_lot = "0.7"
 
 [features]
 default = ["std"]
